+++ src/main.ts
@@ -5,8 +5,5 @@ export { Control } from "./components/Control/Control"
 export { DynamicCard, setDynamicCardDefaults } from "./components/DynamicCard/DynamicCard"
 export { Image } from "./components/Image/Image"
 export { Product } from "./components/Product/Product"
-
-export { SimpleCard, setSimpleCardDefaults } from "./components/SimpleCard/SimpleCard"
 export { SkuOptions } from "./components/SkuOptions/SkuOptions"
-export { VariantSelector } from "./components/VariantSelector/VariantSelector"
 export { setShopifyShop } from "./shopify/getShopifyUrl"