+++ test/setup.ts
@@ -6,10 +6,7 @@ import "@/components/Control/Control"
 import "@/components/DynamicCard/DynamicCard"
 import "@/components/Image/Image"
 import "@/components/Product/Product"
-
-import "@/components/SimpleCard/SimpleCard"
 import "@/components/SkuOptions/SkuOptions"
-import "@/components/VariantSelector/VariantSelector"
 import "@/components/Bundle/Bundle"
 
 beforeAll(() => {