+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### BREAKING CHANGES
 
 * **SectionCampaign:** Remove SectionCampaign component
+* **SimpleCard:** Remove SimpleCard component
+* **VariantSelector:** Remove VariantSelector component
 
 ## [10.18.9](https://github.com/Nosto/web-components/compare/v10.18.8...v10.18.9) (2026-02-04)